batch script for executing validator added

+++ run.sh
@@ -1,5 +1,22 @@
-cd ./result
-rm *.csv
+#!/bin/bash
+
+cd result
+if [ -f "*.csv" ]; then
+   rm *.csv
+fi
+
+if [ -d "pool" ]; then
+   echo "pool already exists"
+else
+   mkdir pool
+fi
+
+cd pool
+if [ -d "jsoup" ]; then
+   echo "jsoup already exists"
+else
+   git clone https://github.com/jhy/jsoup
+fi
+
 cd ..
-# python3 ./main.py -g gumtreeVector.csv -t testVector.csv -r 60
-python3 ./main.py -g Grasscutter_vector.csv -c Grasscutter_file_commit.csv -t testVector.csv
+python main.py -g jsoup_gumtree_vector.csv -c jsoup_commit_file.csv -t testVector.csv
